Split serdes
