+++ crates/agentgateway/Cargo.toml
@@ -15,7 +15,7 @@ mimalloc = ["pprof-alloc/allocator-mimalloc", "dep:mimalloc"]
 tls-aws-lc = ["rustls/aws_lc_rs", "tokio-rustls/aws_lc_rs"]
 tls-openssl = ["dep:openssl", "dep:rustls-openssl"]
 ui = []
-schema = ["schemars", "agent-llm/schema"]
+schema = ["schemars", "agent-core/schema", "agent-llm/schema"]
 internal_benches = ["divan"]
 assert_size_runtime = ["agent-core/assert_size_runtime"]
 
@@ -80,7 +80,6 @@ ipnet.workspace = true
 itertools.workspace = true
 jsonwebtoken.workspace = true
 lazy_static.workspace = true
-macro_rules_attribute.workspace = true
 notify-debouncer-full.workspace = true
 notify.workspace = true
 num_cpus.workspace = true
@@ -122,10 +121,8 @@ rustls-pki-types.workspace = true
 rustls.workspace = true
 schemars = { workspace = true, optional = true }
 secrecy.workspace = true
-serde-transcode.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-serde_json_path_to_error.workspace = true
 serde_regex.workspace = true
 serde_urlencoded.workspace = true
 serde_with.workspace = true
