+++ crates/core/src/serdes.rs
@@ -1,0 +1,508 @@
+use std::borrow::Cow;
+use std::fmt::{Debug, Display};
+use std::io;
+use std::path::PathBuf;
+
+use secrecy::SecretString;
+use serde::de::DeserializeOwned;
+use serde::ser::SerializeSeq;
+use serde::{Deserialize, Deserializer, Serializer};
+
+#[cfg(feature = "schema")]
+pub use schemars::JsonSchema;
+#[cfg(not(feature = "schema"))]
+pub trait JsonSchema {}
+
+pub use macro_rules_attribute::{apply, attribute_alias};
+pub use serde_with;
+
+#[macro_export]
+macro_rules! define_schema_aliases {
+	() => {
+		$crate::serdes::attribute_alias! {
+			#[apply(schema_de!)] = #[serde_with::serde_as] #[derive(Debug, Clone, serde::Deserialize)] #[serde(rename_all = "camelCase", deny_unknown_fields)] #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))];
+			#[apply(schema_ser!)] = #[serde_with::serde_as] #[derive(Debug, Clone, serde::Serialize)] #[serde(rename_all = "camelCase", deny_unknown_fields)];
+			#[apply(schema_ser_schema!)] = #[serde_with::serde_as] #[derive(Debug, Clone, serde::Serialize)] #[serde(rename_all = "camelCase", deny_unknown_fields)] #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))];
+			#[apply(schema!)] = #[serde_with::serde_as] #[derive(Debug, Clone, serde::Deserialize, serde::Serialize)] #[serde(rename_all = "camelCase", deny_unknown_fields)] #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))];
+			#[apply(schema_enum!)] = #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, Copy, serde::Deserialize, serde::Serialize)] #[serde(rename_all = "camelCase", deny_unknown_fields)] #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))];
+		}
+	};
+}
+pub use crate::define_schema_aliases;
+
+pub fn is_default<T: Default + PartialEq>(t: &T) -> bool {
+	*t == Default::default()
+}
+
+/// Serde yaml represents things different than just as "JSON in YAML format".
+/// We don't want this. Instead, we transcode YAML via the JSON module.
+pub mod yamlviajson {
+	use serde::{de, ser};
+
+	pub fn from_str<T>(s: &str) -> anyhow::Result<T>
+	where
+		T: for<'de> de::Deserialize<'de>,
+	{
+		let de_yaml = serde_yaml::Deserializer::from_str(s);
+		let mut buf = Vec::with_capacity(128);
+		{
+			let mut se_json = serde_json::Serializer::new(&mut buf);
+			serde_transcode::transcode(de_yaml, &mut se_json)?;
+		}
+		Ok(serde_json_path_to_error::from_slice(&buf)?)
+	}
+
+	pub fn to_string<T>(value: &T) -> anyhow::Result<String>
+	where
+		T: ?Sized + ser::Serialize,
+	{
+		let js = serde_json::to_string(value)?;
+		let mut buf = Vec::with_capacity(128);
+		let mut se_yaml = serde_yaml::Serializer::new(&mut buf);
+		let de_serde = serde_yaml::Deserializer::from_str(&js);
+		serde_transcode::transcode(de_serde, &mut se_yaml)?;
+		Ok(String::from_utf8(buf)?)
+	}
+}
+
+pub struct SerAsStr;
+impl<T> serde_with::SerializeAs<T> for SerAsStr
+where
+	T: AsRef<str>,
+{
+	fn serialize_as<S>(source: &T, serializer: S) -> Result<S::Ok, S::Error>
+	where
+		S: Serializer,
+	{
+		serializer.serialize_str(source.as_ref())
+	}
+}
+impl<'de, T> serde_with::DeserializeAs<'de, T> for SerAsStr
+where
+	T: std::str::FromStr,
+	<T as std::str::FromStr>::Err: Display,
+{
+	fn deserialize_as<D>(deserializer: D) -> Result<T, D::Error>
+	where
+		D: Deserializer<'de>,
+	{
+		let s = String::deserialize(deserializer)?;
+		s.parse().map_err(serde::de::Error::custom)
+	}
+}
+
+pub mod serde_instant_option {
+	use std::time::{Duration, Instant};
+
+	use crate::durfmt;
+	use crate::prelude::AtomicOption;
+	use serde::Serializer;
+
+	pub fn serialize<S: Serializer>(
+		t: &AtomicOption<Instant>,
+		serializer: S,
+	) -> Result<S::Ok, S::Error> {
+		match t.load().as_ref() {
+			None => serializer.serialize_none(),
+			Some(t) => {
+				let ts: Duration = t
+					.checked_duration_since(Instant::now())
+					.unwrap_or(Duration::ZERO);
+				serializer.serialize_str(durfmt::format(ts).as_str())
+			},
+		}
+	}
+}
+
+pub mod serde_dur {
+	use std::time::Duration;
+
+	use crate::durfmt;
+	use serde::{Deserialize, Deserializer, Serializer};
+
+	pub fn serialize<S: Serializer>(t: &Duration, serializer: S) -> Result<S::Ok, S::Error> {
+		serializer.serialize_str(&durfmt::format(*t))
+	}
+
+	pub fn deserialize<'de, D>(deserializer: D) -> Result<Duration, D::Error>
+	where
+		D: Deserializer<'de>,
+	{
+		let input = String::deserialize(deserializer)?;
+
+		durfmt::parse(&input)
+			.map_err(|e| serde::de::Error::custom(format!("failed to parse duration: {e:?}")))
+	}
+}
+
+pub mod serde_dur_option {
+	use std::time::Duration;
+
+	use crate::durfmt;
+	use serde::{Deserialize, Deserializer, Serializer};
+
+	pub fn serialize<S: Serializer>(t: &Option<Duration>, serializer: S) -> Result<S::Ok, S::Error> {
+		match t {
+			None => serializer.serialize_none(),
+			Some(t) => serializer.serialize_str(&durfmt::format(*t)),
+		}
+	}
+
+	pub fn deserialize<'de, D>(deserializer: D) -> Result<Option<Duration>, D::Error>
+	where
+		D: Deserializer<'de>,
+	{
+		let input: Option<String> = Option::deserialize(deserializer)?;
+
+		input
+			.as_deref()
+			.map(durfmt::parse)
+			.transpose()
+			.map_err(serde::de::Error::custom)
+	}
+}
+
+pub mod serde_base64 {
+	use base64::Engine;
+	use base64::prelude::BASE64_STANDARD;
+	use serde::{Deserialize, Deserializer, Serializer};
+
+	pub fn serialize<T, S>(key: &T, serializer: S) -> Result<S::Ok, S::Error>
+	where
+		T: AsRef<[u8]>,
+		S: Serializer,
+	{
+		serializer.serialize_str(&BASE64_STANDARD.encode(key.as_ref()))
+	}
+
+	pub fn deserialize<'de, T, D>(deserializer: D) -> Result<T, D::Error>
+	where
+		D: Deserializer<'de>,
+		T: From<Vec<u8>>,
+	{
+		use serde::de::Error;
+		String::deserialize(deserializer)
+			.and_then(|string| {
+				BASE64_STANDARD
+					.decode(&string)
+					.map_err(|err| Error::custom(err.to_string()))
+			})
+			.map(|bytes| T::from(bytes))
+	}
+}
+
+pub fn ser_display_option<S: Serializer, T: Display>(
+	t: &Option<T>,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	match t {
+		None => serializer.serialize_none(),
+		Some(t) => serializer.serialize_str(&t.to_string()),
+	}
+}
+
+pub fn serde_scheme_opt<S: Serializer>(
+	t: &Option<&http::uri::Scheme>,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	match t {
+		None => serializer.serialize_none(),
+		Some(t) => serializer.serialize_str(t.as_str()),
+	}
+}
+
+pub fn serde_authority_opt<S: Serializer>(
+	t: &Option<&http::uri::Authority>,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	match t {
+		None => serializer.serialize_none(),
+		Some(t) => serializer.serialize_str(t.as_str()),
+	}
+}
+
+pub fn ser_display_iter<S: Serializer, T, TI: Display>(
+	t: &T,
+	serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+	for<'a> &'a T: IntoIterator<Item = &'a TI>,
+{
+	let mut seq = serializer.serialize_seq(None)?;
+	for el in t {
+		seq.serialize_element(&el.to_string())?;
+	}
+	seq.end()
+}
+
+pub fn ser_display<S: Serializer, T: Display>(t: &T, serializer: S) -> Result<S::Ok, S::Error> {
+	serializer.serialize_str(&t.to_string())
+}
+
+pub fn ser_debug<S: Serializer, T: Debug>(t: &T, serializer: S) -> Result<S::Ok, S::Error> {
+	serializer.serialize_str(&format!("{t:?}"))
+}
+
+pub fn ser_redact<S: Serializer, T>(_: &T, serializer: S) -> Result<S::Ok, S::Error> {
+	serializer.serialize_str("<redacted>")
+}
+
+fn is_sensitive_header(name: &str) -> bool {
+	name.eq_ignore_ascii_case("authorization")
+}
+
+pub fn ser_sensitive_header_map<S: Serializer>(
+	map: &std::collections::HashMap<String, String>,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	use serde::ser::SerializeMap;
+	let mut m = serializer.serialize_map(Some(map.len()))?;
+	for (k, v) in map {
+		let v: &str = if is_sensitive_header(k) {
+			"<redacted>"
+		} else {
+			v
+		};
+		m.serialize_entry(k, v)?;
+	}
+	m.end()
+}
+
+pub fn ser_sensitive_header_vec<S: Serializer>(
+	vec: &Vec<(String, String)>,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	use serde::ser::SerializeSeq;
+	let mut s = serializer.serialize_seq(Some(vec.len()))?;
+	for (k, v) in vec {
+		let v: &str = if is_sensitive_header(k) {
+			"<redacted>"
+		} else {
+			v
+		};
+		s.serialize_element(&(k, v))?;
+	}
+	s.end()
+}
+
+pub fn ser_string_or_bytes<S: Serializer, T: AsRef<[u8]>>(
+	t: &T,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	let b = t.as_ref();
+	if let Ok(s) = std::str::from_utf8(b) {
+		serializer.serialize_str(s)
+	} else {
+		serializer.serialize_bytes(b)
+	}
+}
+
+pub fn ser_string_or_bytes_option<S: Serializer, T: AsRef<[u8]>>(
+	t: &Option<T>,
+	serializer: S,
+) -> Result<S::Ok, S::Error> {
+	match t {
+		None => serializer.serialize_none(),
+		Some(t) => ser_string_or_bytes(t, serializer),
+	}
+}
+
+pub fn ser_bytes<S: Serializer, T: AsRef<[u8]>>(t: &T, serializer: S) -> Result<S::Ok, S::Error> {
+	let b = t.as_ref();
+	if let Ok(s) = std::str::from_utf8(b) {
+		serializer.serialize_str(s)
+	} else {
+		serializer.serialize_str(&hex::encode(b))
+	}
+}
+
+pub fn de_parse<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+where
+	D: Deserializer<'de>,
+	for<'a> T: TryFrom<&'a str>,
+	for<'a> <T as TryFrom<&'a str>>::Error: Display,
+{
+	let s: Cow<'de, str> = Cow::<str>::deserialize(deserializer)?;
+	match T::try_from(s.as_ref()) {
+		Ok(t) => Ok(t),
+		Err(e) => Err(serde::de::Error::custom(e)),
+	}
+}
+
+pub fn de_parse_option<'de, D, T>(deserializer: D) -> Result<Option<T>, D::Error>
+where
+	D: Deserializer<'de>,
+	for<'a> T: TryFrom<&'a str>,
+	for<'a> <T as TryFrom<&'a str>>::Error: Display,
+{
+	let s: Option<Cow<'de, str>> = Option::deserialize(deserializer)?;
+	let Some(s) = s else { return Ok(None) };
+	match T::try_from(s.as_ref()) {
+		Ok(t) => Ok(Some(t)),
+		Err(e) => Err(serde::de::Error::custom(e)),
+	}
+}
+
+pub fn de_bytes<S: Serializer, T: AsRef<[u8]>>(t: &T, serializer: S) -> Result<S::Ok, S::Error> {
+	let b = t.as_ref();
+	if let Ok(s) = std::str::from_utf8(b) {
+		serializer.serialize_str(s)
+	} else {
+		serializer.serialize_str(&hex::encode(b))
+	}
+}
+
+pub fn deser_key_from_file<'de, D>(deserializer: D) -> Result<SecretString, D::Error>
+where
+	D: Deserializer<'de>,
+{
+	let input = FileOrInline::deserialize(deserializer)?;
+	load_secret(input).map_err(serde::de::Error::custom)
+}
+
+pub fn deser_key_from_file_option<'de, D>(deserializer: D) -> Result<Option<SecretString>, D::Error>
+where
+	D: Deserializer<'de>,
+{
+	let input = Option::<FileOrInline>::deserialize(deserializer)?;
+	let Some(input) = input else {
+		return Ok(None);
+	};
+	load_secret(input)
+		.map(Some)
+		.map_err(serde::de::Error::custom)
+}
+
+fn load_secret(input: FileOrInline) -> Result<SecretString, String> {
+	let key = input.load().map_err(|e| e.to_string())?;
+	Ok(SecretString::from(key.trim().to_string()))
+}
+
+pub fn deser_key<'de, D>(deserializer: D) -> Result<SecretString, D::Error>
+where
+	D: Deserializer<'de>,
+{
+	let input = String::deserialize(deserializer)?;
+
+	Ok(SecretString::from(input))
+}
+
+pub fn de_as<'de, I, O, D>(deserializer: D) -> Result<O, D::Error>
+where
+	D: Deserializer<'de>,
+	I: DeserializeOwned,
+	O: TryFrom<I>,
+	<O as TryFrom<I>>::Error: Display,
+{
+	let s: I = I::deserialize(deserializer)?;
+	O::try_from(s).map_err(serde::de::Error::custom)
+}
+
+pub fn de_as_opt<'de, I, O, D>(deserializer: D) -> Result<Option<O>, D::Error>
+where
+	D: Deserializer<'de>,
+	I: DeserializeOwned,
+	O: TryFrom<I>,
+	<O as TryFrom<I>>::Error: Display,
+{
+	let s: Option<I> = <Option<I>>::deserialize(deserializer)?;
+	match s {
+		Some(i) => Ok(Some(O::try_from(i).map_err(serde::de::Error::custom)?)),
+		None => Ok(None),
+	}
+}
+
+#[derive(Debug, Clone, serde::Deserialize)]
+#[cfg_attr(feature = "schema", derive(JsonSchema))]
+#[serde(untagged)]
+pub enum FileOrInline {
+	File { file: PathBuf },
+	Inline(String),
+}
+
+impl FileOrInline {
+	pub fn load(&self) -> io::Result<String> {
+		match self {
+			FileOrInline::File { file } => fs_err::read_to_string(file),
+			FileOrInline::Inline(s) => Ok(s.clone()),
+		}
+	}
+}
+
+#[derive(Clone, Default, Debug)]
+pub struct RenamedField;
+
+pub fn renamed_field<'de, D>(old: &'static str, new: &'static str, _: D) -> Result<(), D::Error>
+where
+	D: Deserializer<'de>,
+{
+	Err(serde::de::Error::custom(format!(
+		"`{}` has been removed; move to `{}`",
+		old, new
+	)))
+}
+
+pub trait ConstString: Default {
+	const VALUE: &str;
+	fn as_str(&self) -> &'static str {
+		Self::VALUE
+	}
+}
+
+#[macro_export]
+macro_rules! const_string {
+	($name:ident = $value:literal) => {
+		#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
+		pub struct $name;
+
+		impl $crate::serdes::ConstString for $name {
+			const VALUE: &str = $value;
+		}
+
+		impl serde::Serialize for $name {
+			fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+			where
+				S: serde::Serializer,
+			{
+				serializer.serialize_str($value)
+			}
+		}
+
+		impl<'de> serde::Deserialize<'de> for $name {
+			fn deserialize<D>(deserializer: D) -> Result<$name, D::Error>
+			where
+				D: serde::Deserializer<'de>,
+			{
+				let s: String = serde::Deserialize::deserialize(deserializer)?;
+				if s == $value {
+					Ok($name)
+				} else {
+					Err(serde::de::Error::custom(format!(concat!(
+						"expect const string value \"",
+						$value,
+						"\""
+					))))
+				}
+			}
+		}
+
+		#[cfg(feature = "schemars")]
+		impl schemars::JsonSchema for $name {
+			fn schema_name() -> std::borrow::Cow<'static, str> {
+				std::borrow::Cow::Borrowed(stringify!($name))
+			}
+
+			fn json_schema(_: &mut schemars::SchemaGenerator) -> schemars::Schema {
+				use serde_json::{Map, json};
+
+				let mut schema_map = Map::new();
+				schema_map.insert("type".to_string(), json!("string"));
+				schema_map.insert("format".to_string(), json!("const"));
+				schema_map.insert("const".to_string(), json!($value));
+
+				schemars::Schema::from(schema_map)
+			}
+		}
+	};
+}
+pub use crate::const_string;
